backend/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
aiosqlite
openai
anthropic
google-generativeai
cryptography
python-dotenv
pydantic
httpx
